requirements.txt
fastapi[standard]
bcrypt
newsapi-python